+++ CompGeoAlgoVisualizer/Source/UI/IDropDown.h
@@ -12,16 +12,16 @@ namespace D2D
 class IDropDown : public IContainer
 {
 	private:
-		const int PADDING = 5;
+		const float PADDING = 5.0f;
 		class IDropDownOption : public IComponent
 		{
 			friend IDropDown;
 			public:
 				IDropDownOption(IDropDown* pDropDown, std::wstring& string, int index);
-				void setWidth(int width) override;
-				void setHeight(int height) override;
-				void onPress(int x, int y, MouseEvent& e) override;
-				void onHover(int x, int y, MouseEvent& e) override;
+				void setWidth(float width) override;
+				void setHeight(float height) override;
+				void onPress(float x, float y, MouseEvent& e) override;
+				void onHover(float x, float y, MouseEvent& e) override;
 				void onExit() override;
 			private:
 				IDropDown* pDropDown;
@@ -32,18 +32,18 @@ class IDropDown : public IContainer
 		};
 	friend IDropDownOption;
 	public:
-		IDropDown(std::wstring text, int width, int height);
-		IDropDown(std::wstring text, D2D::TextFormat& textFormat, int width, int height);
-		IDropDown(int width, int height);
-		IDropDown(D2D::TextFormat& textFormat, int width, int height);
+		IDropDown(std::wstring text, float width, float height);
+		IDropDown(std::wstring text, D2D::TextFormat& textFormat, float width, float height);
+		IDropDown(float width, float height);
+		IDropDown(D2D::TextFormat& textFormat, float width, float height);
 		~IDropDown();
 		void setColor(const Color& color) override;
 		int getSelectedIndex();
-		void setWidth(int width) override;
-		void setHeight(int height) override;
+		void setWidth(float width) override;
+		void setHeight(float height) override;
 		void addOption(std::wstring text);
 		void addChild(IComponent* child) override;
-		void onPress(int x ,int y, MouseEvent& e) override;
+		void onPress(float x ,float y, MouseEvent& e) override;
 		void onUpdate(IComponent* parent) override;
 		void onFocusLoss() override;
 	private:
