+++ CompGeoAlgoVisualizer/Source/UI/ILabel.cpp
@@ -4,7 +4,7 @@
 
 using namespace D2D;
 
-ILabel::ILabel(std::wstring text, TextFormat& textFormat, int width, int height)
+ILabel::ILabel(std::wstring text, TextFormat& textFormat, float width, float height)
 	:
 	IComponent(0, 0, width, height),
 	pText(new Text(text, textFormat, width, height))
@@ -12,13 +12,13 @@ ILabel::ILabel(std::wstring text, TextFormat& textFormat, int width, int height)
 	mDrawables.push_back(pText);
 }
 
-void ILabel::setWidth(int width)
+void ILabel::setWidth(float width)
 {
 	IComponent::setWidth(width);
 	pText->setWidth(width);
 }
 
-void ILabel::setHeight(int height)
+void ILabel::setHeight(float height)
 {
 	IComponent::setHeight(height);
 	pText->setHeight(height);
