+++ CompGeoAlgoVisualizer/Source/UI/IPanel.cpp
@@ -2,7 +2,7 @@
 #include "IPanel.h"
 #include "Direct2D/Drawable/Rectangle.h"
 
-IPanel::IPanel(int width, int height, Color color)
+IPanel::IPanel(float width, float height, Color color)
 	:
 	IContainer(0, 0, width, height),
 	pRect(new D2D::Rectangle({ 0, 0 }, width, height, true, color))
@@ -10,13 +10,13 @@ IPanel::IPanel(int width, int height, Color color)
 	mDrawables.push_back(pRect);
 }
 
-void IPanel::setWidth(int width)
+void IPanel::setWidth(float width)
 {
 	mWidth = width;
 	pRect->setWidth(width);
 }
 
-void IPanel::setHeight(int height)
+void IPanel::setHeight(float height)
 {
 	mHeight = height;
 	pRect->setHeight(height);
